MapTopologia/requirements.txt
customtkinter==5.2.0
Pillow==10.0.1
scapy==2.5.0
networkx==3.1
matplotlib==3.7.2
reportlab==4.0.4
python-nmap==0.7.1